+++ src/Service/LogicalOperations/LogicalOperations.java
@@ -1,5 +1,5 @@
 package Service.LogicalOperations;
 
 public interface LogicalOperations {
-    int[] doLogicalOperations(int[] a, int[] b);
+    byte[] doLogicalOperations(int[] a, int[] b);
 }
